Add Func() flag
This implements the `Func()` flag as seen in Go 1.16. This
allows one to define a `Func` flag which takes the usual name and
usage, as well as a `func(string) error` which is used as a callback.

The callback is called every time the flag is called allowing for custom
code to manage how the flag value is set.

Fix https://github.com/spf13/pflag/issues/335