Remove comma separation for flags
The CSV parsing, as well as the normal comma separation, is
extremely inconsistent, and tends to mess with input (e.g. #307).

This commit removes all CSV related code in flag argument parsing
and also removes the `StringArray()` as the code of it is now redundant
since `StringSlice()` does the same with the CSV logic removed.

Closes #307