Allow values of `StringTo*` to be optional
This allows the values of `StringTo*` to be optional. This that, when
the option has been enabled, the value `--flag name` is valid. This
allows for the `StringTo*` to act like a `StringSlice`.

When retrieving, this because the key of within the `map[string]*`
result.

The tests for these types have also been improved.